Engines QT - STM
Transmito informacion proveniente de un slider en QT hacia STM

+++ Core/Inc/ADC.h
@@ -18,7 +18,7 @@
  * @param whiteValue:   Valor que corresponde a valores white
  */
 typedef struct{
-	volatile uint16_t bufferADCvalue[NUM_CHANNELS_ADC]; /*!< Es volatile porque lo llena el DMA */
+	volatile uint32_t bufferADCvalue[NUM_CHANNELS_ADC]; /*!< Es volatile porque lo llena el DMA */
 	uint16_t currentValue[NUM_CHANNELS_ADC];
 }_sIrSensor;
